backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
openai>=1.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
